# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AppendZipResponse:
    properties:
      added:
        type: integer
      collection_id:
        type: integer
      cover_key:
        type: string
      file_count:
        type: integer
      prefix:
        type: string
      source_zip_key:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for Emoji collection site
  title: Emoji API
  version: 0.1.0
paths:
...
  /api/admin/collections/{id}/import-zip:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: collection id
        in: path
        name: id
        required: true
        type: integer
      - description: zip file
        in: formData
        name: file
        required: true
        type: file
      - description: set first file as cover
        in: formData
        name: set_cover
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AppendZipResponse'
      summary: Append zip to existing collection
      tags:
      - admin
...
swagger: "2.0"